students/K33401/Kiriltsev_Roman/laboratory_work_2/untitled1/requirements.txt
cryptography==3.1.1
django-allauth